+++ src/app/Http/Controllers/Admin/UserManagement/UserController.php
@@ -31,6 +31,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize(User::class);
+
         return view('admin.user_management.users.create', [
             'user' => [],
         ]);
@@ -44,6 +46,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize(User::class);
+
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
@@ -78,6 +82,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize($user);
+
         return view('admin.user_management.users.edit', [
             'user' => $user
         ]);
@@ -92,6 +98,8 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        $this->authorize($user);
+
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', Rule::unique('users')->ignore($user->id)],
@@ -114,6 +122,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize($user);
+
         $user->delete();
 
         return redirect()->route('admin.user_management.user.index');
